backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
pydantic>=2.7.0
pypdf>=4.2.0
python-docx>=1.1.2
llama-cpp-python>=0.2.77
openai>=1.30.0  # <--- NEW